+++ NeAccounting/DomainShared/ViewModels/Workers/FunctionViewModel.cs
@@ -32,7 +32,7 @@ namespace DomainShared.ViewModels.Workers
         /// <summary>
         /// ماه مربوط به کارکرد
         /// </summary>
-        public int PersianMonth { get; set; }
+        public byte PersianMonth { get; set; }
 
         /// <summary>
         /// سال مربوط به کارکرد
